requirements.txt
torch
torchvision
timm
numpy
pillow
opencv-python
facenet-pytorch
scikit-learn
matplotlib
tqdm
